--- apps/watch/src/components/Hero/index.ts
@@ -1,1 +1,0 @@
-export { VideoHero } from './VideoHero/VideoHero'
