+++ apps/watch/src/components/VideoContentPage/VideoHero/VideoHero.tsx
@@ -1,0 +1,135 @@
+import Box from '@mui/material/Box'
+import CircularProgress from '@mui/material/CircularProgress'
+import Stack from '@mui/material/Stack'
+import Typography from '@mui/material/Typography'
+import { ReactElement, useRef, useState } from 'react'
+import { secondsToMinutes } from '@core/shared/ui/timeFormat'
+import Container from '@mui/material/Container'
+import Button from '@mui/material/Button'
+import PlayArrow from '@mui/icons-material/PlayArrow'
+import AccessTime from '@mui/icons-material/AccessTime'
+import Circle from '@mui/icons-material/Circle'
+import { useVideo } from '../../../libs/videoContext'
+import { VideoHeroPlayer } from './VideoHeroPlayer'
+
+import 'video.js/dist/video-js.css'
+
+interface VideoHeroProps {
+  loading?: boolean
+}
+
+export function VideoHero({ loading }: VideoHeroProps): ReactElement {
+  const { image, variant, children, title } = useVideo()
+  const [isPlaying, setIsPlaying] = useState(false)
+  const videoRef = useRef<HTMLVideoElement>(null)
+
+  function playVideo(): void {
+    setIsPlaying(true)
+    if (videoRef?.current != null) {
+      videoRef?.current?.play()
+    }
+  }
+
+  return (
+    <>
+      {loading === true && <CircularProgress />}
+      <>
+        <Box
+          sx={{
+            backgroundImage: `url(${image as string})`,
+            backgroundSize: 'cover',
+            backgroundPosition: 'center',
+            height: 776
+          }}
+        >
+          <VideoHeroPlayer videoRef={videoRef} playVideo={playVideo} />
+          {!isPlaying && (
+            <>
+              <Container
+                maxWidth="xl"
+                style={{
+                  position: 'absolute',
+                  top: 350,
+                  paddingLeft: 100,
+                  margin: 0,
+                  textShadow: '0px 3px 4px rgba(0, 0, 0, 0.25)'
+                }}
+              >
+                <Typography
+                  variant="h2"
+                  sx={{
+                    maxWidth: '600px',
+                    color: 'text.primary'
+                  }}
+                >
+                  {title[0]?.value}
+                </Typography>
+              </Container>
+              <Box
+                sx={{
+                  position: 'absolute',
+                  top: '520px'
+                }}
+                width="100%"
+                height="133px"
+              >
+                <Stack
+                  direction="row"
+                  justifyContent="space-between"
+                  px="100px"
+                  sx={{ color: 'text.primary' }}
+                >
+                  <Stack direction="row" spacing="20px">
+                    {children.length > 0 && (
+                      <Typography variant="subtitle1">
+                        {children.length} episodes
+                      </Typography>
+                    )}
+                    {children.length === 0 && (
+                      <>
+                        <Button
+                          size="large"
+                          variant="contained"
+                          sx={{ height: 71, fontSize: '24px' }}
+                          onClick={playVideo}
+                        >
+                          <PlayArrow />
+                          &nbsp; Play Video
+                        </Button>
+                        {variant !== null && (
+                          <Stack height="71px" direction="row">
+                            <AccessTime sx={{ paddingTop: '23px' }} />
+                            <Typography
+                              variant="body2"
+                              sx={{ lineHeight: '71px', paddingLeft: '10px' }}
+                            >
+                              {secondsToMinutes(variant.duration)} min
+                            </Typography>
+                          </Stack>
+                        )}
+                        <Circle sx={{ fontSize: '10px', paddingTop: '30px' }} />
+                      </>
+                    )}
+                  </Stack>
+                </Stack>
+              </Box>
+              <Box
+                sx={{
+                  backgroundColor: 'rgba(18, 17, 17, 0.25)',
+                  position: 'absolute',
+                  top: '643px'
+                }}
+                width="100%"
+                height="133px"
+              >
+                <Stack pt="34px" mx="100px" width="100%" direction="row">
+                  <Stack direction="row">&nbsp;</Stack>
+                </Stack>
+              </Box>
+            </>
+          )}
+        </Box>
+      </>
+    </>
+  )
+}
