+++ apps/watch/src/components/VideoContentPage/VideoHero/index.ts
